+++ src/math/equation.cpp
@@ -67,7 +67,7 @@ class MyProgram : public Program
 			else {
 				eqsolver::variable_print vprinter(dim.height, dim.width, x-1 ,y-1, &grid[internal], NULL);
 				eqsolver::ast_print<eqsolver::variable_print> solver(&vprinter);
-				result[internal] = solver(ast);
+				result[internal] = (int)solver(ast);
 			}
 		}
